neuroglancer/neuro-demo/server/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
cloud-volume
pyspng-seunglab
python-multipart
tifffile
